+++ src/views/MainView.vue
@@ -254,7 +254,7 @@ let filteredStory = computed(() =>
         <ul tabindex="0" class="dropdown-content menu p-2 rounded-box shadow-xl bg-primary mt-4">
           <li class="p-1"><label for="setting-modal">Setting</label></li>
           <li class="p-1"><label for="about-modal">About</label></li>
-          <li v-if="installPromptMain"><label @click="installPromptMain.prompt()">Install</label></li>
+          <li class="p-1" v-if="installPromptMain"><label @click="installPromptMain.prompt()">Install</label></li>
         </ul>
       </div>
     </div>
@@ -405,13 +405,5 @@ let filteredStory = computed(() =>
       <StorageDetail />
     </template>
   </ModalComponent>
-
-  <!-- <input type="checkbox" id="check-modal" class="modal-toggle" />
-  <div :class="['modal modal-bottom sm:modal-middle ', storageStatus ? 'modal-open' : '']">
-    <div class="modal-box relative flex flex-col items-center">
-      <h3 class="font-bold text-lg mb-4 text-center">Please Give Me A Moment. Checking LocalStorage Space. It Takes Up To 20 Sec. This Message Will Close After It's Done.</h3>
-      <progress class="progress w-56"></progress>
-    </div>
-  </div> -->
 
 </template>